moved site specific config file into config directory
